✨ feat: Refactor project structure and implement new vector store services
