+++ test_requirements.txt
@@ -1,3 +1,5 @@
 pytest==8.3.4
 pytest-asyncio==0.26.0
 pytest-postgresql==7.0.1
+mongomock==4.3.0
+httpx==0.27.0
